マイグレーションファイルを修正2

+++ db/migrate/20250604060231_remove_play_tickets_and_edit_tickets_from_users.rb
@@ -1,6 +1,6 @@
 class RemovePlayTicketsAndEditTicketsFromUsers < ActiveRecord::Migration[7.0]
   def change
-    remove_column :users, :play_tickets, :integer
-    remove_column :users, :edit_tickets, :integer
+    remove_column :users, :play_tickets, :integer if column_exists?(:users, :play_tickets)
+    remove_column :users, :edit_tickets, :integer if column_exists?(:users, :edit_tickets)
   end
 end
